[LINT] linted + utils file for general data manipulation.
