+++ .gitignore
@@ -3,4 +3,5 @@
 data/*
 maps/*
 .vscode/
+__pycache__/
 ########################
