+++ src/cities_processing.py
@@ -1,32 +1,10 @@
+"""Module for gathering lists of cities,
+and output to data/cities/cities_processing_output.json"""
 import json
 from tqdm import tqdm
 from shapely.geometry import shape
+from utils import create_polygon, reduce_multi_shape, reduce_shape
 
-
-def reduce_shape(geometry):
-    """Reduce the number of points of the given shape
-    Returns: a list of coordinates
-    """
-    coords = None
-    ratio = 0.001
-
-    coords = geometry.simplify(ratio, preserve_topology=True)
-
-    res = []
-
-    for coord in coords.exterior.coords:
-        res.append([coord[1], coord[0]])
-    return res
-
-
-def reduce_multi_shape(geometry):
-    """Applies the reduce shape function on a MultiPolygon
-    Returns: a list a reduced shape
-    """
-    res = []
-    for geom in geometry.geoms:
-        res.append(reduce_shape(geom))
-    return res
 
 def init_coordonates(geometry):
     """Initialize the new coordinates and the list of points to visit
@@ -34,17 +12,19 @@ def init_coordonates(geometry):
     """
     coord = geometry["coordinates"]
     if geometry["type"] == "MultiPolygon":
-        liste_a_parcourir = []
-        for k in coord:
-            liste_a_parcourir += k[0]
-            points_cardinaux = [coord[0][0][0][0], coord[0][0][0][0],
-                               coord[0][0][0][1], coord[0][0][0][1]]
-        return liste_a_parcourir, points_cardinaux
+        list_to_iterate = []
 
-    liste_a_parcourir = coord[0]
-    points_cardinaux = [coord[0][0][0], coord[0][0][0],
+        for k in coord:
+            list_to_iterate += k[0]
+
+        cardinal_points = [coord[0][0][0][0], coord[0][0][0][0],
+                           coord[0][0][0][1], coord[0][0][0][1]]
+        return list_to_iterate, cardinal_points
+
+    list_to_iterate = coord[0]
+    cardinal_points = [coord[0][0][0], coord[0][0][0],
                        coord[0][0][1], coord[0][0][1]]
-    return liste_a_parcourir, points_cardinaux
+    return list_to_iterate, cardinal_points
 
 
 
@@ -64,50 +44,31 @@ def create_cities_file():
         cities_list.append([i["properties"], i["geometry"]])
 
     print("Writing file...")
-    with open('data/cities/cities_output_Valentin.json', 'w',encoding='utf-8') as json_file:
+    with open('data/cities/cities_processing_output.json', 'w',encoding='utf-8') as json_file:
         json_file.write("[\n")
-        compteur = 1
+        counter = 1
         p_bar = tqdm(total=len(cities_list))
+
         for [prop, geom] in cities_list:
-            liste_a_parcourir, points_cardinaux = init_coordonates(geom)
+            liste_a_parcourir, cardinal_points = init_coordonates(geom)
 
             for k in liste_a_parcourir:
-                if k[0] < points_cardinaux[0]:
-                    points_cardinaux[0] = k[0]
-                if k[0] > points_cardinaux[1]:
-                    points_cardinaux[1] = k[0]
-                if k[1] < points_cardinaux[2]:
-                    points_cardinaux[2] = k[1]
-                if k[1] > points_cardinaux[3]:
-                    points_cardinaux[3] = k[1]
+                if k[0] < cardinal_points[0]:
+                    cardinal_points[0] = k[0]
+                if k[0] > cardinal_points[1]:
+                    cardinal_points[1] = k[0]
+                if k[1] < cardinal_points[2]:
+                    cardinal_points[2] = k[1]
+                if k[1] > cardinal_points[3]:
+                    cardinal_points[3] = k[1]
 
             if prop["code"][:2] == "2A" or prop["code"][:2] == "2B":
                 dep = 20
             else:
                 dep = int(prop["code"][:2])
 
-            polygon = {
-                "type": "Feature",
-                        "geometry": {
-                            "type": "Polygon",
-                            "coordinates": [
-                                    [
-                                        [points_cardinaux[1],
-                                         points_cardinaux[3]],
-                                        [points_cardinaux[0],
-                                         points_cardinaux[3]],
-                                        [points_cardinaux[0],
-                                         points_cardinaux[2]],
-                                        [points_cardinaux[1],
-                                         points_cardinaux[2]]
-                                    ]
-                                ],
-                            },
-                        "properties": {
-                            "nom": prop["nom"],
-                            "dep": dep
-                        }
-                    }
+
+            polygon = create_polygon(cardinal_points, prop["nom"], dep)
 
             if geom["type"] == 'MultiPolygon':
                 polygon['geometry']['MultiShape'] = [reduce_multi_shape(shape(geom))]
@@ -116,8 +77,8 @@ def create_cities_file():
                 polygon['geometry']['shape'] = [reduce_shape(shape(geom))]
 
             json.dump(polygon, json_file, ensure_ascii=False)
-            compteur += 1
-            if compteur != len(cities_list) + 1:
+            counter += 1
+            if counter != len(cities_list) + 1:
                 json_file.write(",\n")
             p_bar.update(1)
 
@@ -127,6 +88,5 @@ def create_cities_file():
     file.close()
     print("End of execution: no error")
     return True
-
 
 create_cities_file()
